Refactor: Product quantity

+++ product-service/src/main/java/com/ttdat/productservice/application/commands/product/CreateProductCommand.java
@@ -22,7 +22,7 @@ public class CreateProductCommand {
 
     String description;
 
-    Integer totalQuantity;
+    Double totalQuantity;
 
     String imageUrl;
